refactor: split lib/types.ts into domain-specific type files with backwards-compat re-exports
Split 1301-line frontend/src/lib/types.ts into 7 focused domain modules
under frontend/src/types/ (core, library, wanted, translation, providers,
settings, system). The original lib/types.ts is replaced with a barrel
re-export so all existing imports continue to work unchanged.
